Moving/consolidating bin to top of the package
This will allow much more flexibility as we move forward and we can avoid breaking changes as we work to improve the build tools.

Updates to bin
-  Now accepts a project as an argument
-  Provides error messaging when a project is not included.
