+++ package.json
@@ -7,12 +7,10 @@
     "node": ">=6"
   },
   "files": [
-    "extensions/pocket-new-tab/bin",
-    "extensions/save-to-pocket/bin"
+    "bin"
   ],
   "bin": {
-    "build-pocket-new-tab": "./extensions/pocket-new-tab/bin/build.js",
-    "build-save-to-pocket": "./extensions/save-to-pocket/bin/build.js"
+    "build-pocket": "./bin/run.js"
   },
   "repository": {
     "type": "git",
